+++ tests/.cache/v/cache/lastfailed
@@ -1,0 +1,5 @@
+{
+  "test_dojo.py": true,
+  "test_person.py": true,
+  "test_rooms.py": true
+}
