+++ tests/test_person.py
@@ -3,8 +3,8 @@ Test cases for the commandline program 'create_room.py'
 '''
 
 import unittest
-from personel.person import Person, Staff, Fellow
-from rooms.room import Office, LivingSpace, Room
+from person.person import Person, Staff, Fellow
+from room.room import Office, LivingSpace, Room
 from dojo_manager import DojoManager
 
 
@@ -24,13 +24,20 @@ class TestAddPerson(unittest.TestCase):
         self.assertIsInstance(daniel, Fellow, 'Formed object not of the '
                                              'required class')
 
+    def test_wrong_argument(self):
+        my_class = Staff()
+        jeff = my_class.add_person({['Daniel', 'Kitui'], 'M'})
+
 
 class Test_add_person_to_a_room(unittest.TestCase):
     def test_add_to_existing_room(self):
         office_instance = Office()
         new_office = office_instance.create_room('Kilimanjaro', 'Office')
         new_person = Staff().add_person(['Steve', 'Man'])
-        DojoManager.allocate_office(['Steve', 'Man'])
+        name = ['Steve', 'Man']
+        DojoManager.allocate_office(name)
+
+
 
 if __name__ == '__main__':
     unittest.main()
